# cmd/server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_uniedit_server_internal_module_payment_provider.PaymentScene:
    enum:
    - web
    - h5
    - app
    - native
    - mini
    type: string
    x-enum-comments:
      PaymentSceneApp: Native app payment
      PaymentSceneH5: Mobile web payment
      PaymentSceneMini: Mini program payment
      PaymentSceneNative: QR code / scan payment
      PaymentSceneWeb: Desktop web payment
    x-enum-descriptions:
    - Desktop web payment
    - Mobile web payment
    - Native app payment
    - QR code / scan payment
    - Mini program payment
    x-enum-varnames:
    - PaymentSceneWeb
    - PaymentSceneH5
    - PaymentSceneApp
    - PaymentSceneNative
    - PaymentSceneMini
...
  internal_module_payment.CreateNativePaymentRequest:
    properties:
      method:
        allOf:
        - $ref: '#/definitions/internal_module_payment.PaymentMethod'
        enum:
        - alipay
        - wechat
      openid:
        description: Required for WeChat mini/jsapi
        type: string
      order_id:
        type: string
      return_url:
        type: string
      scene:
        allOf:
        - $ref: '#/definitions/github_com_uniedit_server_internal_module_payment_provider.PaymentScene'
        enum:
        - web
        - h5
        - app
        - native
        - mini
    required:
    - method
    - order_id
    - scene
    type: object
  internal_module_payment.CreatePaymentIntentRequest:
    properties:
      order_id:
        type: string
    required:
    - order_id
    type: object
  internal_module_payment.NativePaymentResponse:
    properties:
      amount:
        type: integer
      app_pay_data:
        description: App SDK payment data
        type: string
      currency:
        type: string
      expire_time:
        description: Unix timestamp
        type: integer
      method:
        $ref: '#/definitions/internal_module_payment.PaymentMethod'
      mini_pay_data:
        additionalProperties:
          type: string
        description: Mini program payment params
        type: object
      order_id:
        type: string
      pay_url:
        description: Redirect URL for web/h5
        type: string
      payment_id:
        type: string
      qr_code:
        description: QR code content for native
        type: string
    type: object
  internal_module_payment.PaymentIntentResponse:
    properties:
      amount:
        type: integer
      client_secret:
        type: string
      currency:
        type: string
      payment_intent_id:
        type: string
    type: object
  internal_module_payment.PaymentMethod:
    enum:
    - card
    - alipay
    - wechat
    type: string
    x-enum-varnames:
    - PaymentMethodCard
    - PaymentMethodAlipay
    - PaymentMethodWechat
  internal_module_payment.PaymentResponse:
    properties:
      amount:
        type: integer
      created_at:
        type: string
      currency:
        type: string
      failed_at:
        type: string
      failure_code:
        type: string
      failure_message:
        type: string
      id:
        type: string
      method:
        $ref: '#/definitions/internal_module_payment.PaymentMethod'
      order_id:
        type: string
      provider:
        type: string
      refunded_amount:
        type: integer
      status:
        $ref: '#/definitions/internal_module_payment.PaymentStatus'
      succeeded_at:
        type: string
    type: object
  internal_module_payment.PaymentStatus:
    enum:
    - pending
    - processing
    - succeeded
    - failed
    - canceled
    - refunded
    type: string
    x-enum-varnames:
    - PaymentStatusPending
    - PaymentStatusProcessing
    - PaymentStatusSucceeded
    - PaymentStatusFailed
    - PaymentStatusCanceled
    - PaymentStatusRefunded
...
host: localhost:8080
info:
  contact:
    email: support@uniedit.io
    name: UniEdit Support
    url: https://uniedit.io/support
  description: UniEdit 视频编辑器后端 API
  license:
    name: Proprietary
    url: https://uniedit.io/license
  termsOfService: https://uniedit.io/terms
  title: UniEdit Server API
  version: "1.0"
paths:
...
  /payments/{id}:
    get:
      description: Get details of a specific payment
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_module_payment.PaymentResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get payment
      tags:
      - Payment
  /payments/intent:
    post:
      consumes:
      - application/json
      description: Create a Stripe payment intent for an order
      parameters:
      - description: Payment intent request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_module_payment.CreatePaymentIntentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_module_payment.PaymentIntentResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create payment intent
      tags:
      - Payment
  /payments/methods:
    get:
      description: Get all saved payment methods for the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List payment methods
      tags:
      - Payment
  /payments/native:
    post:
      consumes:
      - application/json
      description: Create an Alipay or WeChat Pay payment for an order
      parameters:
      - description: Native payment request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_module_payment.CreateNativePaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_module_payment.NativePaymentResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create native payment
      tags:
      - Payment
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer token authentication. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"